+++ app/Http/Controllers/Admin/TypologyController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use App\Order;
+use App\Http\Controllers\Controller;
+use App\Typology;
 use Illuminate\Http\Request;
 
-class OrderController extends Controller
+class TypologyController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -41,10 +42,10 @@ class OrderController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Order  $order
+     * @param  \App\Typology  $typology
      * @return \Illuminate\Http\Response
      */
-    public function show(Order $order)
+    public function show(Typology $typology)
     {
         //
     }
@@ -52,10 +53,10 @@ class OrderController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Order  $order
+     * @param  \App\Typology  $typology
      * @return \Illuminate\Http\Response
      */
-    public function edit(Order $order)
+    public function edit(Typology $typology)
     {
         //
     }
@@ -64,10 +65,10 @@ class OrderController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Order  $order
+     * @param  \App\Typology  $typology
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Order $order)
+    public function update(Request $request, Typology $typology)
     {
         //
     }
@@ -75,10 +76,10 @@ class OrderController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Order  $order
+     * @param  \App\Typology  $typology
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Order $order)
+    public function destroy(Typology $typology)
     {
         //
     }
